Add a chance to dim at maximum allowed by device
There are tablets which limit by hardware the minimal screen brightness (0 doesn't mean completely off). This is a chance to reach this level as most as possible. 100% doesn't work, but with 99% device can be dimmed to the lowest possible value allowed by the unit. This is desirable in cases when the manufacturer already thougt about prohibiting a completely black screen by hardware design.